requirements.txt
hydra-core==1.3.2
numpy==1.23.3
omegaconf==2.2.3
pandas==1.5.1
torch==1.10.1+cu111
torchaudio==0.10.1+cu111
tqdm==4.64.1
